Adjust benchmark runner to allow parallel runs
The new -j argument will parallelize the benchmark run
with however many processes are specified. Note that the
benchmark script runs in a separate thread, so the -j
argument should be one less than the number of cores
you want to use.

Each process is pinned to a single CPU using the CPU
affinity flag. The script itself is always pinned to
CPU0.